requirements.txt
rasa==3.6.20
newsapi-python
openai==1.33.0
python-dotenv==1.0.1